@@ -4,7 +4,7 @@ version: 0.0.1
 homepage: https://github.com/ened/flutter_plugin_auth0_native.git
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.13+hotfix.6 <2.0.0"
 
 dependencies:
